requirements.txt
fastapi
uvicorn[standard]
websockets
twilio
python-dotenv
audioop-lts
numpy
httpx
kokoro>=0.9.2
